refactor: updated market home from sliders to buttons
